moved transaction status to separate component
